refactor tests and types
