+++ server/package.json
@@ -19,11 +19,11 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.7",
-    "@types/jest": "^26.0.4",
+    "@types/jest": "^26.0.14",
     "@types/mongoose": "^5.7.31",
     "@types/node": "^14.0.22",
     "@types/supertest": "^2.0.10",
-    "jest": "^26.1.0",
+    "jest": "^26.6.0",
     "nodemon": "^2.0.4",
     "supertest": "^4.0.2",
     "ts-jest": "^26.1.2",
