+++ server/src/utils/shortid.ts
@@ -1,4 +1,4 @@
-import { nanoid } from 'nanoid'
+import { nanoid } from "nanoid"
 
 export function generate() {
   return nanoid(8)
