--- server/tests/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "sourceMap": false,
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["../src/*"]
-    }
-  },
-  "include": ["./**/*.ts"],
-  "exclude": ["node_modules"]
-}
